requirements.txt
aiosqlite==0.17.0; python_version >= "3.7" and python_version < "4.0"
anyio==3.3.0; python_full_version >= "3.6.2" and python_version >= "3.6"
asyncpg==0.24.0; python_full_version >= "3.6.0"
atomicwrites==1.4.0; python_version >= "3.6" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.6" and python_full_version >= "3.4.0"
attrs==21.2.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
certifi==2021.5.30; python_version >= "3.6"
charset-normalizer==2.0.4; python_full_version >= "3.5.0" and python_version >= "3.6"
colorama==0.4.4; python_version >= "3.6" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.6" and python_full_version >= "3.5.0"
fastapi==0.68.1; python_version >= "3.6"
h11==0.12.0; python_version >= "3.6"
httpcore==0.13.6; python_version >= "3.6"
httpx==0.19.0; python_version >= "3.6"
idna==3.2; python_version >= "3.6" and python_full_version >= "3.6.2"
iniconfig==1.1.1; python_version >= "3.6"
iso8601==0.1.16; python_version >= "3.7" and python_version < "4.0"
packaging==21.0; python_version >= "3.6"
pluggy==1.0.0; python_version >= "3.6"
py==1.10.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
pydantic==1.8.2; python_full_version >= "3.6.1" and python_version >= "3.6"
pyparsing==2.4.7; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
pypika-tortoise==0.1.1; python_version >= "3.7" and python_version < "4.0"
pytest-asyncio==0.15.1; python_version >= "3.6"
pytest==6.2.5; python_version >= "3.6"
python-dotenv==0.19.0; python_version >= "3.5"
pytz==2021.1; python_version >= "3.7" and python_version < "4.0"
rfc3986==1.5.0; python_version >= "3.6"
sniffio==1.2.0; python_full_version >= "3.6.2" and python_version >= "3.6"
starlette==0.14.2; python_version >= "3.6"
toml==0.10.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
tortoise-orm==0.17.7; python_version >= "3.7" and python_version < "4.0"
typing-extensions==3.10.0.2; python_full_version >= "3.6.1" and python_version >= "3.7" and python_version < "4.0"
